requirements.txt
asgiref==3.8.1
Django==5.2
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
greenlet==3.1.1
gunicorn==23.0.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
psycopg==3.2.6
psycopg2==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.40
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
